Deprecate old network module
