+++ rust/src/network_deprecated/buffer_utils.rs
