+++ rust/src/network_deprecated/data_writer.rs
