+++ rust/src/network_deprecated/mod.rs
@@ -4,7 +4,6 @@ pub mod sockets;
 pub mod data_writer;
 pub mod data_reader;
 pub mod utils;
-pub mod py_interface;
 pub mod buffer_utils;
 pub mod buffer_queues;
 pub mod remote_transfer_handler;
@@ -12,3 +11,4 @@ pub mod metrics;
 pub mod network_config;
 pub mod sockets_monitor;
 pub mod channels_router;
+pub mod py_interface;
