+++ rust/src/network_deprecated/utils.rs
