+++ rust/src/newtork/mod.rs
@@ -10,3 +10,4 @@ pub mod data_reader;
 pub mod metrics;
 pub mod network_config;
 pub mod remote_transfer_handler;
+pub mod py_interface;
