+++ rust/src/newtork/remote_transfer_handler.rs
@@ -5,9 +5,7 @@ use crossbeam::{channel::{bounded, Receiver, Select, Sender}, queue::ArrayQueue}
 use pyo3::{pyclass, pymethods};
 use serde::{Deserialize, Serialize};
 
-use crate::network::metrics::NUM_BYTES_RECVD;
-
-use super::{buffer_utils::{get_channeld_id, Bytes}, channel::{to_local_and_remote, Channel}, metrics::{MetricsRecorder, NUM_BUFFERS_RECVD, NUM_BUFFERS_SENT, NUM_BYTES_SENT}, io_loop::{IOHandler, CROSSBEAM_DEFAULT_CHANNEL_SIZE}, sockets::{channels_to_socket_identities, parse_ipc_path_from_addr, SocketIdentityGenerator, SocketKind, SocketMetadata}};
+use super::{buffer_utils::{get_channeld_id, Bytes}, channel::{to_local_and_remote, Channel}, io_loop::{IOHandler, CROSSBEAM_DEFAULT_CHANNEL_SIZE}, metrics::{MetricsRecorder, NUM_BUFFERS_RECVD, NUM_BUFFERS_SENT, NUM_BYTES_RECVD, NUM_BYTES_SENT}, sockets::{channels_to_socket_identities, parse_ipc_path_from_addr, SocketIdentityGenerator, SocketKind, SocketMetadata}};
 
 #[derive(Serialize, Deserialize, Clone)]
 #[pyclass(name="RustTransferConfig")]
